Rough dashboard implementation, authenticated layout change and added new header layout
