+++ app/Http/Controllers/ShowCourseLessonsController.php
@@ -9,7 +9,7 @@ use Illuminate\Http\Request;
 
 class ShowCourseLessonsController extends Controller
 {
-    public function __invoke($course_id): Response {
+    public function __invoke($course_id) {
         $lessons = Lesson::all()->where('course_id', $course_id)->toArray();
         // dd($lessons);
         return Inertia('Lessons', [
