+++ resources/js/Pages/Dashboard.vue
@@ -1,26 +1,14 @@
 <script setup>
-import AuthenticatedLayout from '@/Layouts/AuthenticatedLayout.vue';
-import { Head } from '@inertiajs/inertia-vue3';
+    import AuthenticatedLayout from '@/Layouts/AuthenticatedLayout.vue';
+    import Header from '@/Layouts/Header.vue';
+    import { Link } from '@inertiajs/inertia-vue3';
+    import GuestLayout from '@/Layouts/GuestLayout.vue';
+    import Lessons from '@/Components/Lessons.vue';
+    const props = defineProps(['lessons']);
 </script>
 
 <template>
-    <Head title="Dashboard" />
-
-    <AuthenticatedLayout>
-        <template #header>
-            <h2 class="font-semibold text-xl text-gray-800 leading-tight">
-                Dashboard
-            </h2>
-        </template>
-
-        <div class="py-12">
-            <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
-                <div class="bg-white overflow-hidden shadow-sm sm:rounded-lg">
-                    <div class="p-6 bg-white border-b border-gray-200">
-                        You're logged in!
-                    </div>
-                </div>
-            </div>
-        </div>
-    </AuthenticatedLayout>
+    <Header>
+        <AuthenticatedLayout :lessons="lessons"/>
+    </Header>
 </template>
